Move the trait functions/definitions to a dedicated file
