+++ frontend/exporter/src/types/mod.rs
@@ -4,6 +4,7 @@ mod mir;
 mod new;
 mod replaced;
 mod todo;
+mod traits;
 
 pub use copied::*;
 pub use index::*;
@@ -11,3 +12,4 @@ pub use mir::*;
 pub use new::*;
 pub use replaced::*;
 pub use todo::*;
+pub use traits::*;
